Fix function name conflicts and one bug
1. read_data_serial_h5 conflicts with that in meta_data.c. Changed to
   read_data_h5
2. ratio_emax in get_particle_tags is float, not int. Stupid error
   Why didn't the complier give warning on this?

+++ vpic-sorter/particle_tags.h
@@ -1,2 +1,2 @@
-void get_particle_tags(char *filename, int tstep, int ratio_emax,
+void get_particle_tags(char *filename, int tstep, float ratio_emax,
         int num_ptl, int *tags);
